Extract class DeleteImposterTests
