+++ MbDotNet.Tests/MbDotNet.Tests.csproj
@@ -67,6 +67,7 @@
   <ItemGroup>
     <Compile Include="Client\CreateHttpImposterTests.cs" />
     <Compile Include="Client\CreateTcpImposterTests.cs" />
+    <Compile Include="Client\DeleteImposterTests.cs" />
     <Compile Include="Models\Imposters\HttpsImposterTests.cs" />
     <Compile Include="Models\Imposters\HttpImposterTests.cs" />
     <Compile Include="Models\Imposters\TcpImposterTests.cs" />
